// models/DISPATCH/drivers/driversDetails.ts
export interface DriversDetails {
    id: number;
    carrier_id: number;
    driver_last_name: string;
    updated_date: Date|string;
    driver_first_name: string;
    driver_id: number;
    notes: string;
    driver_type: string;
    acceptance: number;
    driver_avatar: number;
    hire_date: string;
    driver_phone_number: string;
    driver_dob: string;
    driver_email_address: string;
    on_time: number;
    driver_live_location: string;
    driver_score: number;
    payment_rate: number;
    status: string;
}